Set buffer_space to long long incase it becomes negative to prevent wrap around which would result in unwanted behaviour

+++ src/libsac2c/global/ctinfo.c
@@ -194,7 +194,7 @@ Format2Buffer (const char *format, va_list arg_p)
 
     if (len_p >= message_buffer_size) {
         /* buffer too small  */
-        
+
         MEMfree (message_buffer);
         message_buffer = (char *)MEMmalloc (len_p + 2);
         CHKMdoNotReport (message_buffer);
@@ -203,7 +203,7 @@ Format2Buffer (const char *format, va_list arg_p)
         va_copy (arg_p_copy, arg_p);
         len_p = (size_t)(len = vsnprintf (message_buffer, message_buffer_size, format, arg_p_copy));
         va_end (arg_p_copy);
-        
+
 
         DBUG_ASSERT (len >= 0 || len_p < message_buffer_size, "message buffer corruption");
     }
@@ -1367,7 +1367,7 @@ CTIfunParams (node *fundef)
     size_t tmp_str_size;
 
     static char argtype_buffer[80];
-    static size_t buffer_space;
+    static long long buffer_space;
 
     DBUG_ENTER ();
 
@@ -1383,9 +1383,9 @@ CTIfunParams (node *fundef)
             tmp_str = TYtype2String (AVIS_TYPE (ARG_AVIS (arg)), TRUE, 0);
             tmp_str_size = STRlen (tmp_str);
 
-            if ((tmp_str_size + 3) <= buffer_space) {
+            if (((long long)tmp_str_size + 3) <= buffer_space) {
                 strcat (argtype_buffer, tmp_str);
-                buffer_space -= tmp_str_size;
+                buffer_space -=(long long)tmp_str_size;
                 if (ARG_NEXT (arg) != NULL) {
                     strcat (argtype_buffer, ", ");
                     buffer_space -= 2;
